requirements.txt
Pillow==6.1.0
python-telegram-bot
regex
python-gitlab
